refactor: FragmentOptions namespace and improve error handling
Moved FragmentOptions to Nalix.Framework.Options and updated using directives accordingly. Changed FragmentHeader size check to throw InternalErrorException with clearer error messages. This improves code organization and diagnostics.
