+++ src/Nalix.SDK/Transport/Internal/FRAME_READER.cs
@@ -15,6 +15,7 @@ using Nalix.Framework.DataFrames.Chunks;
 using Nalix.Framework.Extensions;
 using Nalix.Framework.Injection;
 using Nalix.Framework.Memory.Buffers;
+using Nalix.Framework.Options;
 using Nalix.SDK.Configuration;
 
 namespace Nalix.SDK.Transport.Internal;
