+++ src/Nalix.SDK/Transport/Internal/FRAME_SENDER.cs
@@ -15,6 +15,7 @@ using Nalix.Framework.DataFrames.Chunks;
 using Nalix.Framework.Injection;
 using Nalix.Framework.Memory.Buffers;
 using Nalix.SDK.Configuration;
+using Nalix.Framework.Options;
 
 namespace Nalix.SDK.Transport.Internal;
